# Conflicts:
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/emotes/EmoteMenuHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/models/DragonModel.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/ClientDragonRenderer.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/ClawsAndTeethRenderLayer.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/DragonArmorRenderLayer.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/DragonGlowLayerRenderer.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/DragonRenderer.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/commands/DragonCommand.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/blocks/DragonDoor.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/blocks/SmallDragonDoor.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/capability/DragonStateHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/capability/DragonStateProvider.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/entity/creatures/Hunter.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/DragonBonusHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/DragonConfigHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/DragonFoodHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/DragonPenaltyHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/HunterOmenHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/magic/ClawToolHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/common/handlers/magic/MagicHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/config/ServerConfig.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/magic/abilities/ForestDragon/active/HunterAbility.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/magic/common/DragonAbility.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/magic/common/passive/AthleticsAbility.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/EntityMixin.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/LivingEntityMixin.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/PlayerMixin.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/network/NetworkHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/network/magic/SyncVisualEffectAdded.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/network/magic/SyncVisualEffectRemoved.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/DSAttributes.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/DSDamageTypes.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/DSEffects.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/DSModifiers.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/DSItemTags.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/DataBlockStateProvider.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/DataBlockTagProvider.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/DataItemModelProvider.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/DataPoiTypeTagsProvider.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/NeoForgedDataGen.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/tags/DSDamageTypeTags.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/datagen/tags/DSEntityTypeTags.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/server/handlers/DragonDestructionHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/server/handlers/EntityHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/server/handlers/PlayerLoginHandler.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/server/handlers/PotionSync.java
#	src/main/java/by/dragonsurvivalteam/dragonsurvival/server/handlers/ServerFlightHandler.java
#	src/main/resources/data/dragonsurvival/enchantment/curse_of_kindness.json
#	src/main/resources/data/dragonsurvival/enchantment/dragonsbane.json
#	src/main/resources/data/dragonsurvival/enchantment/dragonsbonk.json
#	src/main/resources/data/dragonsurvival/enchantment/dragonsboon.json
#	src/main/resources/data/dragonsurvival/tags/item/cold_items.json
#	src/main/resources/data/minecraft/tags/enchantment/tooltip_order.json
#	src/main/resources/data/minecraft/tags/item/chest_armor.json
#	src/main/resources/data/minecraft/tags/item/foot_armor.json
#	src/main/resources/data/minecraft/tags/item/head_armor.json
#	src/main/resources/data/minecraft/tags/item/leg_armor.json

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/server/AmbusherSpawner.java
@@ -72,7 +72,7 @@ public class AmbusherSpawner implements CustomSpawner {
                                     return 0;
                                 } else {
                                     int j1 = 0;
-                                    int k1 = (int)Math.ceil((double)pLevel.getCurrentDifficultyAt(blockpos$mutableblockpos).getEffectiveDifficulty()) + 1;
+                                    int k1 = (int) Math.ceil(pLevel.getCurrentDifficultyAt(blockpos$mutableblockpos).getEffectiveDifficulty()) + 1;
 
                                     for (int l1 = 0; l1 < k1; l1++) {
                                         j1++;
@@ -102,7 +102,7 @@ public class AmbusherSpawner implements CustomSpawner {
     }
 
     private boolean spawnPatrolMember(ServerLevel pLevel, BlockPos pPos, RandomSource pRandom, boolean pLeader) {
-        if(!pLeader) {
+        if (!pLeader) {
             return false;
         }
